feat: move discovery from ui kit to voices
